refactor: create a module and export Run function
